+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this repository will be documented in this file.
 
 
 ## [Unreleased]
+### Added
+- Add early stopping with validation dataset split.
 
 ## [0.0.12] - 2022-1-5
 ### Fixed
@@ -16,6 +18,7 @@ All notable changes to this repository will be documented in this file.
 ### Added
 - Add `class_weights` support
 - Add retry logic when loading images from ibdoc
+- Add options for early stopping and evaluation dataset creation
 ### Fixed
 - Add fix inside data processing to correct bboxes with negative coordinates
 ### Changed